Migrated some calendar related stuff to yourloops
